# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "econpaper-backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sessions/{session_id}/transform": {
      "post": {
        "summary": "Run Transform",
        "description": "Apply variable recoding & construction (sub-step 5).\n\nRequest body (flat form, matching the frontend VariableConstructor):\n``{\"type\": \"log_transform\"|\"onehot\"|\"label\"|\"bin\"|\"interaction\"|\"policy_dummy\",\n   \"column\": str, ...}``\n\nThe flat ``type``/``column`` is translated into the nested config that\n``cleaning.transform.transform`` expects.",
        "operationId": "run_transform_sessions__session_id__transform_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransformRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransformResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/filter": {
      "post": {
        "summary": "Run Filter",
        "description": "Apply sample filtering (sub-step 6).\n\nRequest body: ``{\"conditions\": [{\"col\": str, \"op\": str, \"val\": Any}, ...]}``",
        "operationId": "run_filter_sessions__session_id__filter_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/balance": {
      "post": {
        "summary": "Run Balance",
        "description": "Check panel balance (sub-step 7).\n\nRequest body: ``{\"panel_id\": str, \"time_col\": str}``\n\nStage D 修复漂移 1：返回 ``BalanceResponse``（4 字段：\n``balanced`` / ``unbalanced`` / ``n_periods`` / ``attrition_rate``），\n不再直接透传 step report。``unbalanced`` 在 router 层补算\n（总 panel 数 - balanced 数），因为 BalanceStep 的 step report 不含\n``unbalanced`` 字段（只有 ``balanced`` / ``n_periods`` /\n``attrition_rate``）。",
        "operationId": "run_balance_sessions__session_id__balance_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BalanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BalanceRequest": {
        "properties": {
          "panel_id": {
            "type": "string",
            "title": "Panel Id",
            "default": ""
          },
          "time_col": {
            "type": "string",
            "title": "Time Col",
            "default": ""
          }
        },
        "type": "object",
        "title": "BalanceRequest",
        "description": "POST /sessions/{id}/balance 请求体。"
      },
      "BalanceResponse": {
        "properties": {
          "balanced": {
            "type": "integer",
            "title": "Balanced",
            "default": 0
          },
          "unbalanced": {
            "type": "integer",
            "title": "Unbalanced",
            "default": 0
          },
          "n_periods": {
            "type": "integer",
            "title": "N Periods",
            "default": 0
          },
          "attrition_rate": {
            "type": "number",
            "title": "Attrition Rate",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "BalanceResponse",
        "description": "POST /sessions/{id}/balance 返回体。\n\n修复漂移 1：统一字段名为 ``balanced`` / ``unbalanced`` / ``n_periods``\n/ ``attrition_rate``。后端 BalanceStep 的 step report 只含\n``balanced`` / ``n_periods`` / ``attrition_rate``，``unbalanced`` 在\nrouter 层补算（balanced 减去总 panel 数，或在 facade 层补）。"
      },
...
      "FilterConditionItem": {
        "properties": {
          "col": {
            "type": "string",
            "title": "Col"
          },
          "op": {
            "type": "string",
            "title": "Op"
          },
          "val": {
            "title": "Val"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "col",
          "op"
        ],
        "title": "FilterConditionItem",
        "description": "filter 条件项。"
      },
      "FilterRequest": {
        "properties": {
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/FilterConditionItem"
            },
            "type": "array",
            "title": "Conditions"
          }
        },
        "type": "object",
        "title": "FilterRequest",
        "description": "POST /sessions/{id}/filter 请求体。"
      },
      "FilterResultResponse": {
        "properties": {
          "n_before": {
            "type": "integer",
            "title": "N Before"
          },
          "n_after": {
            "type": "integer",
            "title": "N After"
          },
          "conditions": {
            "items": {},
            "type": "array",
            "title": "Conditions"
          }
        },
        "type": "object",
        "required": [
          "n_before",
          "n_after"
        ],
        "title": "FilterResultResponse",
        "description": "POST /sessions/{id}/filter 返回体。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TransformRequest": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "log_transform"
          },
          "column": {
            "type": "string",
            "title": "Column",
            "default": ""
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "TransformRequest",
        "description": "POST /sessions/{id}/transform 请求体（flat form）。"
      },
      "TransformResponse": {
        "properties": {
          "constructed_vars": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Constructed Vars"
          }
        },
        "type": "object",
        "title": "TransformResponse",
        "description": "POST /sessions/{id}/transform 返回体。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}